Authenticate and authorize in JWT

+++ Controllers/api/v1/posts_api.js
@@ -24,15 +24,19 @@ module.exports.index = async function(req,res){
 module.exports.destroy= async function (req,res) {
     try{
         let post=await Post.findById(req.params.id);
+        if(post.user == req.user.id){
+            post.remove();
 
-        post.remove();
+            await Comment.deleteMany({post:req.params.id});
 
-        await Comment.deleteMany({post:req.params.id});
-
-        return res.json(200,{
-            message: "Post and associated comments deleted successfully"
-        });
-    
+            return res.json(200,{
+                message: "Post and associated comments deleted successfully"
+            });
+        }else{
+            res.json(401,{
+                message:'You cannot delete this post'
+            })
+        }
     }catch(err){
 
         console.log('*******',err);
